+++ modules/cloudfront-cache/main.tf
@@ -8,6 +8,7 @@ resource "aws_cloudfront_distribution" "distribution" {
   is_ipv6_enabled = true
   comment         = var.deployment_name
   price_class     = var.cloudfront_price_class
+  aliases         = var.cloudfront_aliases
 
   dynamic "default_cache_behavior" {
     for_each = [var.cloudfront_default_behavior]
@@ -68,7 +69,10 @@ resource "aws_cloudfront_distribution" "distribution" {
   }
 
   viewer_certificate {
-    cloudfront_default_certificate = true
+    acm_certificate_arn            = var.cloudfront_acm_certificate_arn
+    cloudfront_default_certificate = var.cloudfront_acm_certificate_arn == null
+    minimum_protocol_version       = var.cloudfront_minimum_protocol_version
+    ssl_support_method             = var.cloudfront_acm_certificate_arn != null ? "sni-only" : null
   }
 
   restrictions {